Keep bundled elk-test.el of the load-path
Bundling a third-party library should generally be avoided because the
bundled version might end up being loaded for users who also have a
more recent version installed separately.

But because a user who wants to run the tests might not `elk-test`
installed, it might make sense to bundle a copy anyway. This commit
puts `elk-test.el` in a separate directory which also contains a file
`.nosearch`, which ensures that it is not added to the `load-path`.
That file is now loaded using `load` instead of `require`.